telegram/requirements.txt
python-dotenv
telethon==1.32.1
openai
